マージ処理の追加

+++ src/main/scala/com/jellyfish85/dbaccessor/erd/mainte/tool/bean/MsTablesBean.scala
@@ -7,17 +7,18 @@ class MsTablesBean extends GeneralBean {
   val logicalTableName:  String =  ""
   val physicalTableName: String =  ""
 
-  var	trkmIdAttr:	ColumnAttribute[Int] = new ColumnAttribute()
-  var	tabDefIdAttr:	ColumnAttribute[Int] = new ColumnAttribute()
-  var	tableIdAttr:	ColumnAttribute[Int] = new ColumnAttribute()
-  var	revisionAttr:	ColumnAttribute[Int] = new ColumnAttribute()
+  var	trkmIdAttr:	ColumnAttribute[BigDecimal] = new ColumnAttribute()
+  var	tabDefIdAttr:	ColumnAttribute[BigDecimal] = new ColumnAttribute()
+  var	tableIdAttr:	ColumnAttribute[BigDecimal] = new ColumnAttribute()
+  var	revisionAttr:	ColumnAttribute[BigDecimal] = new ColumnAttribute()
   var	logicalTableTagAttr:	ColumnAttribute[String] = new ColumnAttribute()
   var	logicalTableNameAttr:	ColumnAttribute[String] = new ColumnAttribute()
   var	physicalTableNameAttr:	ColumnAttribute[String] = new ColumnAttribute()
   var	trkmStatusAttr:	ColumnAttribute[String] = new ColumnAttribute()
   var	tableCommentAttr:	ColumnAttribute[String] = new ColumnAttribute()
   var	segoseiCheckStatusAttr:	ColumnAttribute[String] = new ColumnAttribute()
-  var	ticketNumberAttr:	ColumnAttribute[Int] = new ColumnAttribute()
+  var	ticketNumberAttr:	ColumnAttribute[BigDecimal] = new ColumnAttribute()
   var	existsFlgAttr:	ColumnAttribute[String] = new ColumnAttribute()
+
 
 }
